Well there's a hefty speed difference.
Generating and testing 16GB with Godot's way takes 236 seconds here. The way Juniper uses takes 42 seconds to get to the same amount.

+++ include/PractRand/RNGs/other/mult.h
@@ -708,7 +708,7 @@ namespace PractRand {
 				class floatHax64 : public vRNG16 {
 					Uint64 state;
 				public:
-					Uint32 rand();
+					Uint64 rand();
 					Uint16 raw16();
 					std::string get_name() const;
 					void walk_state(StateWalkingObject*);
